+++ src/components/font-weight-control/index.js
@@ -4,7 +4,7 @@
 import { __ } from '@wordpress/i18n';
 import SelectControl from '../select-control';
 import { getWeightOptions } from '../typography-control/utils';
-import { loadFonts } from '../../extensions/text/fonts';
+import { loadFontsInEditor } from '../../extensions/text/fonts';
 
 const FontWeightControl = props => {
 	const { onChange, fontName, fontStyle, fontWeight, breakpoint } = props;
@@ -22,12 +22,7 @@ const FontWeightControl = props => {
 				objFont[fontName].weight = val.toString();
 				if (fontStyle) objFont[fontName].style = fontStyle;
 
-				if (breakpoint === 's' || breakpoint === 'xs') {
-					const iframeEditor = document.querySelector(
-						'iframe[name="editor-canvas"]'
-					);
-					loadFonts(objFont, true, iframeEditor.contentDocument);
-				} else loadFonts(objFont);
+				loadFontsInEditor(breakpoint, objFont);
 			}}
 		/>
 	);
