+++ src/extensions/text/fonts/index.js
@@ -1,3 +1,3 @@
 /* eslint-disable import/prefer-default-export */
-export { default as loadFonts } from './loadFonts';
+export * from './loadFonts';
 export * from './utils';
